Move `captureStream(…)` from `mas` to `masTests`.
Resolve #686
